[FLINK-16316][operators] Move inner CountingClass class out from AbstractStreamOperator
